validator: move ServicesSubnet from validator-tests
ServicesSubnet is a struct used in validator, currently it is not a candidate
for a separate e2etest actor, so this commit moves it into the validator
repository.
